Add files via upload
Locks down a program to a specific user.
Created to install metasploit on ubuntu 18.04 without letting the main user account be able to access the executables, could work for pretty much any program though. 
This is to be a single purpose program, ONLY locking one program to one user in one group.